expanding makefile, commented out some main functions, testing all functions

+++ ft_strlen.c
@@ -6,22 +6,22 @@
 /*   By: mbauer <mbauer@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/03 14:46:23 by mbauer            #+#    #+#             */
-/*   Updated: 2025/07/09 11:05:40 by mbauer           ###   ########.fr       */
+/*   Updated: 2025/07/10 16:13:58 by mbauer           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_len_recursiv(const char *s, int i)
+int	ft_len_recursiv(const char *s, size_t i)
 {
 	if (s[i] == '\0')
 		return (i);
 	return (ft_len_recursiv(s, i + 1));
 }
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s[i] != '\0')
